Restructure the modules splitting them out for reuse
